+++ Runtime/Events/EventsManager.cs
@@ -1,4 +1,5 @@
-﻿using AffiseAttributionLib.Executors;
+﻿#nullable enable
+using AffiseAttributionLib.Executors;
 using AffiseAttributionLib.Usecase;
 
 namespace AffiseAttributionLib.Events
@@ -6,6 +7,7 @@ namespace AffiseAttributionLib.Events
     internal class EventsManager
     {
         private const long TIME_SEND_REPEAT = 15 * 1000;
+        private const long SCHEDULER_SEND_REPEAT = 15 * 1000;
 
         private readonly ISendDataToServerUseCase _sendDataToServerUseCase;
         private readonly IExecutorServiceProvider _executorServiceProvider;
@@ -24,9 +26,9 @@ namespace AffiseAttributionLib.Events
             sendEventsOnStart();
         }
 
-        public void SendEvents(bool withDelay = true)
+        public void SendEvents(bool withDelay = true, bool sendEmpty = true)
         {
-            _sendDataToServerUseCase.Send(withDelay);
+            _sendDataToServerUseCase.Send(withDelay, sendEmpty);
         }
 
         private void sendEventsOnStart()
@@ -35,9 +37,26 @@ namespace AffiseAttributionLib.Events
             SendEvents(false);
             
             //Start timer fo repeat send events
+            StartTimer();
+            
+            //Start Scheduler
+            StartQueue();
+        }
+
+        private void StartTimer()
+        {
             _executorServiceProvider.ExecuteWithDelay(TIME_SEND_REPEAT, () =>
             {
                 SendEvents();
+            });
+        }
+
+        private void StartQueue()
+        {
+            _executorServiceProvider.ExecuteWithDelay(SCHEDULER_SEND_REPEAT, () =>
+            {
+                SendEvents(false, false);
+                StartQueue();
             });
         }
     }