+++ Runtime/Parameters/Parameters.cs
@@ -10,8 +10,10 @@
         public const string AFFISE_EVENT_USER_DATA = "affise_event_user_data";
         public const string AFFISE_EVENT_DATA = "affise_event_data";
         public const string AFFISE_PARAMETERS = "affise_parameters";
-
+        public const string AFFISE_EVENT_ID_INDEX = "affise_event_id_index";
+   
         public const string AFFISE_EVENTS_COUNT = "affise_events_count";
         public const string AFFISE_SDK_EVENTS_COUNT = "affise_sdk_events_count";
+        public const string UUID_INDEX_KEY = "uuid_index";
     }
 }