+++ Runtime/Usecase/IIndexUseCase.cs
@@ -1,0 +1,8 @@
+﻿namespace AffiseAttributionLib.Usecase
+{
+    public interface IIndexUseCase
+    {
+        long GetUuidIndex();
+        long GetAffiseEventIdIndex();
+    }
+}